refactor: Move shareConfigOneAccount function to a new file and clean up tests
- Moved the shareConfigOneAccount function from resource_share_acceptance_test.go and resource_share_account_level_acceptance_test.go to a new file resource_share_config_test.go for better organization.
- Removed commented-out code from the test files to improve readability.
