+++ pkg/testacc/resource_share_account_level_acceptance_test.go
@@ -40,13 +40,3 @@ func TestAcc_Share_2025_07(t *testing.T) {
 		},
 	})
 }
-
-// func shareConfigOneAccount(shareId sdk.AccountObjectIdentifier, comment string, account string) string {
-// 	return fmt.Sprintf(`
-// resource "snowflake_share" "test" {
-// 	name           = "%s"
-// 	comment        = "%s"
-// 	accounts       = ["%s"]
-// }
-// `, shareId.Name(), comment, account)
-// }
